requirements.txt
Python=3.9.12
mysql-connector-python=8.0.30
pandas=1.4.3
matplotlib=3.5.1
numpy=1.21.5
mysql=8.0.29
Docker=20.10.17
apache-airflow=2.3.3
